fix(gateway): keep node-policy approval requester eligible as reviewer
A turn-driven plugin approval (e.g. Codex exec-server launch) carried the
operator connection that submitted the turn as its requester, and delivery
excluded that connection from the reviewer set. With a single connected
Control UI client the only eligible reviewer was excluded, so the approval
was auto-denied as no-route before anyone could see it.

Node-invoke policy approvals are minted gateway-side and never presented
through the requesting client's RPC response, so the carried connection is
provenance, not a presenter to dedupe: stop excluding it. RPC-ingress
approval paths keep their requester exclusion.

Also extracts the shared plugin-policy test harness into
node-invoke-plugin-policy.test-helpers.ts (the test file crossed the
max-lines lint cap).
